+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Fixed
 - Fix nullable type on `MyAccount.idpDomain`
 - Fix nullable type on `Branding.ownConditions`
+- Fix nullable type on `EnvelopeTemplateRecipient.intermediaryName`
 
 ### Changed
 - Added native types to all properties