+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Fix nullable type on `MyAccount.idpDomain`
 - Fix nullable type on `Branding.ownConditions`
 - Fix nullable type on `EnvelopeTemplateRecipient.intermediaryName`
+- Fix incorrect type on `EnvelopeTemplateTag.document` and `EnvelopeTemlateTag.recipient`
 
 ### Changed
 - Added native types to all properties