+++ CHANGELOG.md
@@ -7,12 +7,14 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 - Add `EnvelopeRecipientIdentification.authorizedAt`
+- Add resource `Blame`
 
 ### Fixed
 - Fix nullable type on `MyAccount.idpDomain`
 - Fix nullable type on `Branding.ownConditions`
 - Fix nullable type on `EnvelopeTemplateRecipient.intermediaryName`
 - Fix incorrect type on `EnvelopeTemplateTag.document` and `EnvelopeTemlateTag.recipient`
+- Fix incorrect type on `AccountRequest.createdBy`
 
 ### Changed
 - Added native types to all properties