+++ CHANGELOG.md
@@ -12,7 +12,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `Branding.signerReturnUrl`
 - Add `AccountEndpoint.manageBilling`
 - Add multiple fields for `AccountBilling`
-
+- Add identify field for `AccountBilling`
+- 
 ### Fixed
 - Fix nullable type on `MyAccount.idpDomain`
 - Fix nullable type on `Branding.ownConditions`